Move: 기술 검색 부분 컴포넌트 분리
